Add default dates for new important dates
When a user adds a new important date, it will default to the date which
was last input. The idea is that the deadlines are usually near each
other and so when we're inputting a lot of dates around the same time
(potentially a year ahead of today) this should make things a bit easier.